Conflicts:
	ideca2/ideca2-client/src/main/java/pdev/financialbrains/client/boot/Main.java
	ideca2/ideca2-ejb/src/main/java/pdev/financialbrains/ejb/contracts/IFxoptionCrudServicesLocal.java
	ideca2/ideca2-ejb/src/main/java/pdev/financialbrains/ejb/contracts/IUserCrudServiceLocal.java
	ideca2/ideca2-web/src/main/java/pdev/financialbrains/managedBeans/Test.java
	ideca2/ideca2-web/src/main/webapp/pages/trader/contactUs.xhtml

+++ ideca2/ideca2-ejb/src/main/java/pdev/financialbrains/ejb/contracts/ICapFloorServiceRemote.java
@@ -27,7 +27,7 @@ public interface ICapFloorServiceRemote {
 
 	double calculD1(Double cap_floor, float s, float rfr, Double maturity, Double volatility);
 	
-	double pricingCapFloor(Double amount,String cap_floor, Integer nDays, Integer period, Integer daysYear, Double f_rate, Double strike, Double vol, Double maturity, Double rfr);
+	double pricingCapFloor(Double amount,String cap_floor, Integer nDays, Integer period, Integer daysYear, Double f_rate, Float strike, Float vol, Double maturity, Double rfr);
 	
 	
 }
